items_api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
prometheus-client>=0.17.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-sqlalchemy>=0.41b0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
httpx>=0.25.0
redis>=5.0.0
mcp[fastapi]>=1.0.0
prometheus-fastapi-instrumentator>=6.1.0
python-jose[cryptography]>=3.3.0
